[REFACTOR] #17 테이블과 Column 이름 변경
