+++ app/src/main/java/com/example/android_accountbook_13/data/dto/AccountBookItem.kt
@@ -1,0 +1,7 @@
+package com.example.android_accountbook_13.data.dto
+
+data class AccountBookItem(
+    val history: History,
+    val category: Category,
+    val method: Method,
+)
