+++ app/src/main/java/com/example/android_accountbook_13/data/local/datasource/LocalDataSource.kt
@@ -1,16 +1,15 @@
 package com.example.android_accountbook_13.data.local.datasource
 
 import android.database.Cursor
-import com.example.android_accountbook_13.data.AccountBookItem
-import com.example.android_accountbook_13.data.entity.Category
-import com.example.android_accountbook_13.data.entity.History
-import com.example.android_accountbook_13.data.entity.Method
+import com.example.android_accountbook_13.data.dto.Category
+import com.example.android_accountbook_13.data.dto.History
+import com.example.android_accountbook_13.data.dto.Method
 
 interface LocalDataSource {
     fun getHistory(month: Int): Result<Cursor>
     fun insertHistory(history: History): Result<Unit>
-    fun updateHistory(history:History): Result<Unit>
-    fun deleteHistory(history:History): Result<Unit>
+    fun updateHistory(history: History): Result<Unit>
+    fun deleteHistory(history: History): Result<Unit>
 
     fun getAllCategory(): Result<Cursor>
     fun getCategory(id: Int): Result<Cursor>
