+++ app/src/main/java/com/example/android_accountbook_13/data/local/datasource/LocalDataSourceImpl.kt
@@ -1,10 +1,9 @@
 package com.example.android_accountbook_13.data.local.datasource
 
 import android.database.Cursor
-import com.example.android_accountbook_13.data.AccountBookItem
-import com.example.android_accountbook_13.data.entity.Category
-import com.example.android_accountbook_13.data.entity.History
-import com.example.android_accountbook_13.data.entity.Method
+import com.example.android_accountbook_13.data.dto.Category
+import com.example.android_accountbook_13.data.dto.History
+import com.example.android_accountbook_13.data.dto.Method
 import com.example.android_accountbook_13.data.local.SQLiteOpenHelper
 import javax.inject.Inject
