+++ app/src/main/java/com/example/android_accountbook_13/data/local/repository/category/CategoryRepository.kt
@@ -1,7 +1,7 @@
 package com.example.android_accountbook_13.data.local.repository.category
 
 import android.database.Cursor
-import com.example.android_accountbook_13.data.entity.Category
+import com.example.android_accountbook_13.data.dto.Category
 
 interface CategoryRepository {
     fun getAllCategory(): Result<Cursor>
