+++ app/src/main/java/com/example/android_accountbook_13/presenter/MainActivity.kt
@@ -3,6 +3,7 @@ package com.example.android_accountbook_13.presenter
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import com.example.android_accountbook_13.data.local.SQLiteOpenHelper
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
